fix insert tests

+++ test/integration/builder/inserts.js
@@ -1306,45 +1306,58 @@ module.exports = function(knex) {
     });
 
     if (typeof BigInt !== 'undefined') {
-      describe('#3553 insert with bigint', function() {
-        const tableName = 'bigint_insert_tests';
-        before(async function() {
-          await knex.schema.dropTableIfExists(tableName);
-          await knex.schema.createTable(tableName, (t) => {
-            t.integer('id').primary();
-            t.bigInteger('value');
-            if (knex.client.driverName === 'pg') {
-              t.specificType('values', 'bignumber[]');
-            }
-          });
-        });
+      for (const columnType of ['bigint', 'string']) {
+        describe(`#3553 insert with bigint in column of type ${columnType}`, function() {
+          // see: https://github.com/mapbox/node-sqlite3/issues/922
+          // in short: driver converts all selected BigInts as integers
+          if (knex.client.driverName === 'sqlite3' && columnType === 'bigint')
+            return;
 
-        for (const [id, value] of differentBigInts.entries()) {
-          it('should allow insert with BigInt', async function() {
-            await knex(tableName).insert([{ id, value }]);
-            await expect(knex(tableName).where({ id })).to.eventually.be.equal([
-              {
-                id,
-                value: value.toString(),
-              },
-            ]);
+          const tableName = 'bigint_insert_tests';
+          before(async function() {
+            await knex.schema.dropTableIfExists(tableName);
+            await knex.schema.createTable(tableName, (t) => {
+              t.integer('id').primary();
+              if (columnType === 'bigint') {
+                t.bigInteger('value');
+              } else {
+                t.string('value');
+              }
+              if (knex.client.driverName === 'pg') {
+                t.specificType('values', 'bignumber[]');
+              }
+            });
           });
 
-          if (knex.client.driverName === 'pg') {
-            it('should allow insert array with BigInt', async function() {
-              await knex(tableName).insert([{ id, values: [value, value] }]);
+          for (const [id, value] of differentBigInts.entries()) {
+            it('should allow insert with BigInt', async function() {
+              await knex(tableName).insert([{ id, value }]);
               await expect(
                 knex(tableName).where({ id })
-              ).to.eventually.be.equal([
+              ).to.eventually.be.deep.equal([
                 {
-                  id: id,
-                  values: [value.toString(), value.toString()],
+                  id,
+                  value: value.toString(),
                 },
               ]);
             });
+
+            if (knex.client.driverName === 'pg') {
+              it('should allow insert array with BigInt', async function() {
+                await knex(tableName).insert([{ id, values: [value, value] }]);
+                await expect(
+                  knex(tableName).where({ id })
+                ).to.eventually.be.deep.equal([
+                  {
+                    id: id,
+                    values: [value.toString(), value.toString()],
+                  },
+                ]);
+              });
+            }
           }
-        }
-      });
+        });
+      }
     }
   });
 };
